Slashing protection for blocks (#220)

+++ slashing-protection/src/main/resources/migrations/postgresql/V1__initial.sql
@@ -5,15 +5,15 @@ CREATE TABLE validators (
 );
 CREATE TABLE signed_blocks (
     validator_id INTEGER NOT NULL,
-    slot INTEGER NOT NULL,
+    slot NUMERIC(20) NOT NULL,
     signing_root BYTEA NOT NULL,
     FOREIGN KEY(validator_id) REFERENCES validators(id),
     UNIQUE (validator_id, slot)
 );
 CREATE TABLE signed_attestations (
     validator_id INTEGER,
-    source_epoch INTEGER NOT NULL,
-    target_epoch INTEGER NOT NULL,
+    source_epoch NUMERIC(20) NOT NULL,
+    target_epoch NUMERIC(20) NOT NULL,
     signing_root BYTEA NOT NULL,
     FOREIGN KEY(validator_id) REFERENCES validators(id),
     UNIQUE (validator_id, target_epoch)
